+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 ### Fixed
 - Missing `<title>` on WebUI.
+- Infinite "node-in-node-list" notifications.